# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amazon Competitor Intelligence V2",
    "description": "Amazon competitor research API with capture-backed provenance.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/products/{product_id}/analyze": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze",
        "operationId": "analyze_api_products__product_id__analyze_post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionJobRead"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/analyses/{run_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Analysis Detail",
        "operationId": "analysis_detail_api_analyses__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/analyses/{run_id}/claims": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Analysis Claims",
        "operationId": "analysis_claims_api_analyses__run_id__claims_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisClaimResponse"
                  },
                  "title": "Response Analysis Claims Api Analyses  Run Id  Claims Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisClaimResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "claim_type": {
            "type": "string",
            "title": "Claim Type"
          },
          "claim_text": {
            "type": "string",
            "title": "Claim Text"
          },
          "claim_value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Value"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ClaimEvidenceDetail"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "claim_type",
          "claim_text",
          "claim_value",
          "evidence"
        ],
        "title": "AnalysisClaimResponse"
      },
      "AnalysisResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "input_hash": {
            "type": "string",
            "title": "Input Hash"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/AnalysisClaimResponse"
            },
            "type": "array",
            "title": "Claims"
          },
          "withheld_quantitative_claims": {
            "type": "integer",
            "title": "Withheld Quantitative Claims",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "job_id",
          "input_hash",
          "model",
          "prompt_version",
          "schema_version",
          "status",
          "error_message",
          "created_at",
          "completed_at"
        ],
        "title": "AnalysisResponse"
      },
      "ClaimEvidenceDetail": {
        "properties": {
          "observation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Observation Id"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "product_asin": {
            "type": "string",
            "title": "Product Asin"
          },
          "price_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Text"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "collector": {
            "type": "string",
            "title": "Collector"
          },
          "extraction_method": {
            "type": "string",
            "title": "Extraction Method"
          },
          "evidence_id": {
            "type": "string",
            "title": "Evidence Id"
          },
          "evidence_artifact_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Artifact Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "observation_id",
          "product_id",
          "product_asin",
          "price_text",
          "captured_at",
          "source_url",
          "collector",
          "extraction_method",
          "evidence_id",
          "evidence_artifact_id",
          "role"
        ],
        "title": "ClaimEvidenceDetail"
      },
...
      "CollectionJobRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "product_id",
          "status",
          "progress",
          "error_code",
          "error_message",
          "created_at",
          "started_at",
          "finished_at",
          "result",
          "attempts",
          "max_attempts"
        ],
        "title": "CollectionJobRead",
        "description": "Job status response."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "issues": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "error",
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "ValidationIssue": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue"
      }
    }
  }
}